% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{bandwidth_optimisation}
\alias{bandwidth_optimisation}
\title{Bandwith optimization}
\usage{
bandwidth_optimisation(formula, data, dmat, sequence,
  kernel = "bisquare", adaptive = TRUE)
}
\arguments{
\item{formula}{(desc to come)}

\item{data}{(desc to come)}

\item{dmat}{(desc to come)}

\item{sequence}{(desc to come)}

\item{kernel}{function chosen as follow: gaussian, exponential, bisquare, tricube, boxcar}

\item{adaptive}{(desc to come)}
}
\value{
to define
}
\description{
Desc to come
}
