% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_avg.R
\name{bw.CV.F}
\alias{bw.CV.F}
\title{A function for bandwith selection to calibrate a GWPR model, based on the mean over time of the data.}
\usage{
bw.CV.F(
  formula,
  data,
  index,
  effect = c("individual", "time", "twoways", "nested"),
  model = c("within", "random", "ht", "between", "pooling", "fd"),
  kernel = "bisquare",
  dMat,
  interval
)
}
\arguments{
\item{formula}{Regression model formula : Y ~ X1 + ... + Xk}

\item{data}{dataFrame for the Panel data}

\item{index}{List for the indexes : (c(" ID, Time"))}

\item{effect}{the effects introduced in the model, one of "individual", "time", or "twoways" (see plm::plm)}

\item{model}{one of "pooling", "within", "between", "random", "fd", or "ht" (see plm::plm)}

\item{kernel}{gaussian,exponential, bisquare, tricube, boxcar (see GWmodel::gw.weight)}

\item{dMat}{a distance matrix or vector (Optional parameter, see GWmodel::gw.weight)}

\item{interval}{vector of lowest and highest distances to be used in the bw optimization process}
}
\value{
double
}
\description{
A function for bandwith selection to calibrate a GWPR model, based on the mean over time of the data.
}
\examples{
USStates@data$id <- c(1:length(unique(USStates@data[,"state"])))
data <- merge(USStates@data, Produc, by="state", all=T)
dMat <- GWmodel::gw.dist(sp::coordinates(USStates), p=2, longlat=F)
Equation <- log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp
bwCV.F <-  bw.CV.F(formula=Equation, data=data, index=c("id","year"), effect='individual', model="within", kernel="bisquare", dMat=dMat, interval=c(1500000, 2500000))
}
