% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{compute_dmat}
\alias{compute_dmat}
\title{Compute a numeric distance matrix or vector (from GW tools)}
\usage{
compute_dmat(dataset, p, longlat)
}
\arguments{
\item{dataset}{A numeric matrix of two columns giving the coordinates of the data points}

\item{p}{the power of the Minkowski distance, default is 2, i.e. the Euclidean distance}

\item{longlat}{TRUE, great circle distances will be calculated}
}
\value{
Returns    A numeric distance matrix or vector
}
\description{
Compute a numeric distance matrix or vector (from GW tools)
}
