% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{compute_gpwr_plot}
\alias{compute_gpwr_plot}
\title{Dest to come}
\usage{
compute_gpwr_plot(Dataset, P = c("0.05", "0.01", "0.001"))
}
\arguments{
\item{Dataset}{a spacial polygon data frame}

\item{P}{dest to come}
}
\value{
(desc to come)
}
\description{
Dest to come
}
