% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{compute_gwpr}
\alias{compute_gwpr}
\title{Computation of the GWPR model}
\usage{
compute_gwpr(qx_qy, dmat, bandwidth, kernel, adaptive)
}
\arguments{
\item{dmat}{(desc to come)}

\item{kernel}{(desc to come)}

\item{adaptive}{(desc to come)}

\item{bandwith}{(desc to come)}

\item{QX}{(desc to come)}

\item{QY}{(desc to come)}

\item{n}{(total index[1])}

\item{t}{(total index[2])}
}
\value{
wmat, yHat, Resid, HatMat
}
\description{
Computation of the GWPR model
}
