% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{compute_localR2Mat}
\alias{compute_localR2Mat}
\title{Local R-Squared}
\usage{
compute_localR2Mat(QXQY, GWPR)
}
\arguments{
\item{wmat}{(desc to come)}

\item{Qy}{(desc to come)}

\item{yHat}{(desc to come)}
}
\value{
matrix of ....
}
\description{
Local R-Squared
}
