% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{compute_QX_QY}
\alias{compute_QX_QY}
\title{Desc to come}
\usage{
compute_QX_QY(dataset, formula, index, model, effect)
}
\arguments{
\item{dataset}{(desc to come)}

\item{formula}{(desc to come)}

\item{index}{(desc to come)}

\item{model}{(desc to come)}

\item{effect}{(desc to come)}
}
\value{
QX="matrix", QY="numeric", n="numeric", t="numeric", k="numeric", x = "matrix", y = "integer"
}
\description{
Desc to come
}
