% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{compute_R2}
\alias{compute_R2}
\title{Global R-Squared}
\usage{
compute_R2(QX_QY, GWPR)
}
\arguments{
\item{yHat}{(desc to come)}

\item{Qy}{(desc to come)}
}
\value{
R2          (desc to come)
}
\description{
Global R-Squared
}
