% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{compute_shapefile}
\alias{compute_shapefile}
\title{Compute Spacial Polygon Dataframes}
\usage{
compute_shapefile(SpDF, result_shapefile, result_gwpr, result_QX_QY)
}
\arguments{
\item{SpDF}{a spacial polygon data frame}

\item{Shapefile}{a customized shape file}

\item{GWPR}{class result from gwpr method}

\item{QX_QY}{class result from QX_QY method}
}
\value{
(desc to come)
}
\description{
Compute Spacial Polygon Dataframes
}
