% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{compute_std_errors_T_values}
\alias{compute_std_errors_T_values}
\title{Standard errors and T values (matrices)}
\usage{
compute_std_errors_T_values(QX_QY, GWPR)
}
\arguments{
\item{Resid}{(desc to come)}

\item{HatMat}{(desc to come)}
}
\value{
SesMat, TVsMat   (desc to come)
}
\description{
Standard errors and T values (matrices)
}
