% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{data_preparation}
\alias{data_preparation}
\title{Helper function to calculate the mean for each variable of the formula}
\usage{
data_preparation(dataset, formula, id)
}
\arguments{
\item{dataset}{a data frame}

\item{formula}{a formula}

\item{id}{a column from the dataset to group by}
}
\value{
to define
}
\description{
Helper function to calculate the mean for each variable of the formula
}
