% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{gpr}
\alias{gpr}
\title{Return Linear Models for Panel Data}
\usage{
gpr(formula, data, effect, model, index)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data.frame}

\item{effect}{effect introduced in the model c("individual", "time", "twoways", "nested")}

\item{model}{one of c("pooling", "within", "between", "random", "fd", "ht")}

\item{index}{list of indexes}
}
\value{
to define
}
\description{
Return Linear Models for Panel Data
}
