% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_option.R
\name{gwlpr}
\alias{gwlpr}
\title{Desc to come}
\usage{
gwlpr(SpatialDataFrame, dataset, formula, id, index, sequence,
  kernel = "bisquare", effect = "individual", model = "within",
  adaptive = TRUE)
}
\arguments{
\item{SpatialDataFrame}{(desc to come)}

\item{dataset}{(desc to come)}

\item{formula}{(desc to come)}

\item{id}{(desc to come)}

\item{index}{(desc to come)}

\item{sequence}{(desc to come)}

\item{kernel}{(desc to come)}

\item{effect}{(desc to come)}

\item{model}{(desc to come)}

\item{adaptive}{(desc to come)}
}
\value{
(desc to come)
}
\description{
Desc to come
}
\details{
This function runs a ..... regression of .... and calculates a p-value ....
}
\references{
books, articles, papers to be mentionned ...
}
\author{
Maxime and ....
}
