% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Produc}
\alias{Produc}
\title{Produc}
\format{
A list of <to complete>
\itemize{
\item state
\item year
\item region
\item pcap
\item pcap
\item hwy
\item water
\item util
\item pc
\item gsp
\item emp
\item unemp
}
}
\usage{
Produc
}
\description{
List of observations
}
\keyword{datasets}
