% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwplm.R
\name{gwplm}
\alias{gwplm}
\title{This function implements basic geographically weighted panel regression (GWPR)
The function requires the plm packages
Je pense que ce package est utile seulement si l'on imprime les effets individuels/temporels (indisponible pour le moment)}
\usage{
gwplm(
  SpDF,
  data,
  index,
  formula,
  bw,
  kernel,
  adaptive = F,
  dMat,
  effect = c("individual", "time", "twoways", "nested"),
  model = c("within", "random", "ht", "between", "pooling", "fd")
)
}
\arguments{
\item{SpDF:}{large SpatialPolygonsdataFrame}

\item{data:}{dataFrame for the Panel data}

\item{index:}{List for the indexes : (c(" ID, Time"))}

\item{formula:}{Regression model formula : Y ~ X1 + ... + Xk}

\item{bw:}{bandwidth to be used (see GWmodel::gwr.basic)}

\item{kernel:}{gaussian,exponential, bisquare, tricube, boxcar (see GWmodel::gw.weight)}

\item{adaptive:}{TRUE or FALSE (see GWmodel::gw.weight)}

\item{dMat:}{a distance matrix or vector (Optional parameter, see GWmodel::gw.weight)}

\item{effect:}{the effects introduced in the model, one of "individual", "time", or "twoways" (see plm::plm)}

\item{model:}{one of "pooling", "within", "between", "random", "fd", or "ht" (see plm::plm)}
}
\value{

}
\description{
This function implements basic geographically weighted panel regression (GWPR)
The function requires the plm packages
Je pense que ce package est utile seulement si l'on imprime les effets individuels/temporels (indisponible pour le moment)
}
\examples{
data(USStates)
USStates@data$id <- c(1:length(unique(USStates@data[,"state"])))
data <- merge(USStates@data, Produc, by="state", all=T)
dMat <- GWmodel::gw.dist(sp::coordinates(USStates), p=2, longlat=F)
Equation <- log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp
bwCV.A <- bw.CV.A(formula=Equation, data=data, index=c("id","year"), effect='individual', model="within", kernel="bisquare", dMat=dMat, bws=c(30:40))
result <- gwplm(SpDF=USStates, data=data, index=c("id", "year"),formula=Equation, bw = bwCV.A, kernel="bisquare", adaptive=T, effect="individual", model="within", dMat=dMat)
}
