% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Produc}
\alias{Produc}
\title{Produc}
\format{
A list of 816 observations of 11 variables between 1970 and 1986 for 48 states
\itemize{
\item state {US State}
\item year {year}
\item region {region id}
\item pcap
\item pcap
\item hwy
\item water
\item util
\item pc
\item gsp
\item emp
\item unemp
}
}
\usage{
Produc
}
\description{
List of observations
}
\keyword{datasets}
